Fix OpenMP detection
According to the official gcc documentation[1] in order to use OpenMP
one must specify -fopenmp flag. It is wrong to link with internal
compiler library like -lgomp. This breaks other compilers like lcc.
Note that special care for clang is not needed in case -fopenmp is used
as described above.

Furthermore cmake provides standard FindOpenMP.cmake module, so in order
to check for OpenMP it is sufficient to use find_package(openmp).

[1] https://gcc.gnu.org/onlinedocs/libgomp/Enabling-OpenMP.html